requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
faiss-cpu==1.10.0
filelock==3.18.0
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-core==0.3.51
langchain-text-splitters==0.3.8
langsmith==0.3.25
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
openai==1.70.0
orjson==3.10.16
packaging==24.2
paho-mqtt==1.5.1
pillow==11.1.0
pydantic==2.11.2
pydantic_core==2.33.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.0
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
zstandard==0.23.0
